pyspark/requirements.txt
pyspark
Faker
kafka-python
google-cloud-storage
